pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lexical-sae"
version = "2.0.0"
requires-python = ">=3.10"
dependencies = [
    "captum~=0.7.0",
    "datasets~=3.2.0",
    "lime~=0.2.0",
    "nltk~=3.9.0",
    "numpy~=2.2.0",
    "pyyaml~=6.0",
    "scikit-learn~=1.6.0",
    "scipy~=1.15.0",
    "concept-erasure~=0.3.0",
    "schedulefree~=1.4.0",
    "torch~=2.6.0",
    "transformers~=4.48.0",
    "tqdm~=4.67.0",
]

[project.scripts]
lexical-sae = "cajt.scripts.run:main"

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
markers = ["slow: tests requiring network access (deselect with '-k not slow')"]

[tool.setuptools.packages.find]
include = ["cajt*"]
